[SCB-2838]using micrometer to replace spectator part5: configure meter registry
