+++ foundations/foundation-metrics/src/main/java/org/apache/servicecomb/foundation/metrics/publish/MeasurementNode.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.servicecomb.foundation.metrics.publish.spectator;
+package org.apache.servicecomb.foundation.metrics.publish;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
