+++ foundations/foundation-metrics/src/test/java/org/apache/servicecomb/foundation/metrics/TestMetricsBootstrap.java
@@ -39,7 +39,7 @@ import io.micrometer.core.instrument.MeterRegistry;
 import mockit.Deencapsulation;
 
 public class TestMetricsBootstrap {
-  MetricsBootstrap bootstrap = new MetricsBootstrap();
+  MetricsBootstrap bootstrap;
 
   EventBus eventBus = new EventBus();
 
@@ -47,13 +47,15 @@ public class TestMetricsBootstrap {
 
   @BeforeEach
   public void setUp() {
-    bootstrap.setMetricsInitializers(List.of());
     Mockito.when(environment.getProperty(METRICS_WINDOW_TIME, int.class, DEFAULT_METRICS_WINDOW_TIME))
         .thenReturn(DEFAULT_METRICS_WINDOW_TIME);
     Mockito.when(environment.getProperty(
             CONFIG_LATENCY_DISTRIBUTION_MIN_SCOPE_LEN, int.class, 7))
         .thenReturn(7);
-    bootstrap.setEnvironment(environment);
+
+    MetricsBootstrapConfig config = new MetricsBootstrapConfig(environment);
+    bootstrap = new MetricsBootstrap(config);
+    bootstrap.setMetricsInitializers(List.of());
   }
 
   @Test
